requirements.txt
click==8.0.4
colorama==0.4.4
DateTime==4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
SQLAlchemy==1.4.32
uuid==1.30
Werkzeug==2.0.3
WTForms==3.0.1
zope.interface==5.4.0
